PageRank Comparison

+++ Comparable/src/main/java/com/samy/ucr/project/Comparable/pageRank/Matrix.java
@@ -2,15 +2,15 @@ package com.samy.ucr.project.Comparable.pageRank;
 
 public class Matrix {
 
-	public static float[][] multiply(float[][] m1, float[][] m2) {
+	public static double[][] multiply(double[][] m1, double[][] m2) {
 		if (m1[0].length != m2.length)
 			return null;
 		int row = m1.length;
 		int col = m2[0].length;
-		float[][] matrix = new float[row][col];
+		double[][] matrix = new double[row][col];
 		for (int i = 0; i < row; i++)
 			for (int j = 0; j < col; j++) {
-				float sum = 0.0f;
+				double sum = 0.0f;
 				for (int k = 0; k < m1[0].length; k++)
 					sum += m1[i][k] * m2[k][j];
 				matrix[i][j] = sum;
@@ -18,13 +18,13 @@ public class Matrix {
 		return matrix;
 	}
 
-	public static float[] multiply(float[][] m1, float[] m2) {
+	public static double[] multiply(double[][] m1, double[] m2) {
 		if (m1[0].length != m2.length)
 			return null;
 		int row = m1.length;
-		float[] matrix = new float[row];
+		double[] matrix = new double[row];
 		for (int i = 0; i < row; i++) {
-			float sum = 0.0f;
+			double sum = 0.0f;
 			for (int j = 0; j < m2.length; j++) {
 				sum += m1[i][j] * m2[j];
 			}
@@ -33,47 +33,48 @@ public class Matrix {
 		return matrix;
 	}
 
-	public static float[] multiply(float[] m1, float a) {
+	public static double[] multiply(double[] m1, double a) {
 		int row = m1.length;
-		float[] matrix = new float[row];
+		double[] matrix = new double[row];
 		for (int i = 0; i < row; i++)
 			matrix[i] = m1[i] * a;
 		return matrix;
 	}
 
-	public static float[][] minus(float[][] m1, float[][] m2) {
+	public static double[][] minus(double[][] m1, double[][] m2) {
 		if (m1.length != m2.length || m1[0].length != m2[0].length)
 			return null;
 		int row = m1.length;
 		int col = m1[0].length;
-		float[][] matrix = new float[row][col];
+		double[][] matrix = new double[row][col];
 		for (int i = 0; i < row; i++)
 			for (int j = 0; j < col; j++)
 				matrix[i][j] = m1[i][j] - m2[i][j];
 		return matrix;
 	}
 
-	public static float[] minus(float[] m1, float[] m2) {
+	public static double[] minus(double[] m1, double[] m2) {
 		if (m1.length != m2.length)
 			return null;
 		int row = m1.length;
-		float[] matrix = new float[row];
+		double[] matrix = new double[row];
 		for (int i = 0; i < row; i++)
 			matrix[i] = m1[i] - m2[i];
 		return matrix;
 	}
 
-	public static float[] add(float[] m1, float[] m2) {
+	public static double[] add(double[] m1, double[] m2) {
 		if (m1.length != m2.length)
 			return null;
 		int row = m1.length;
-		float[] matrix = new float[row];
+		double[] matrix = new double[row];
 		for (int i = 0; i < row; i++)
 			matrix[i] = m1[i] + m2[i];
 		return matrix;
 	}
 
-	public static boolean absoluteLessThanThreshold(float[][] m1, float threshold) {
+	public static boolean absoluteLessThanThreshold(double[][] m1,
+			double threshold) {
 		int row = m1.length;
 		int col = m1[0].length;
 		for (int i = 0; i < row; i++)
@@ -83,7 +84,8 @@ public class Matrix {
 		return true;
 	}
 
-	public static boolean absoluteLessThanThreshold(float[] m1, float threshold) {
+	public static boolean absoluteLessThanThreshold(double[] m1,
+			double threshold) {
 		int row = m1.length;
 		for (int i = 0; i < row; i++)
 			if (Math.abs(m1[i]) > threshold)
@@ -91,7 +93,7 @@ public class Matrix {
 		return true;
 	}
 
-	public static void printMatrix(float[][] m1) {
+	public static void printMatrix(double[][] m1) {
 		int row = m1.length;
 		int col = m1[0].length;
 		for (int i = 0; i < row; i++) {
@@ -111,7 +113,7 @@ public class Matrix {
 		}
 	}
 
-	public static void printMatrix(float[] m1) {
+	public static void printMatrix(double[] m1) {
 		int row = m1.length;
 		for (int i = 0; i < row; i++)
 			System.out.print(m1[i] + " ");
